Little refactoring MenuFactory
